+++ src/ProxyClient/Noop.php
@@ -12,17 +12,21 @@
 namespace FOS\HttpCache\ProxyClient;
 
 use FOS\HttpCache\ProxyClient\Invalidation\BanCapable;
+use FOS\HttpCache\ProxyClient\Invalidation\ClearCapable;
 use FOS\HttpCache\ProxyClient\Invalidation\PurgeCapable;
 use FOS\HttpCache\ProxyClient\Invalidation\RefreshCapable;
 use FOS\HttpCache\ProxyClient\Invalidation\TagCapable;
 
 /**
- * This is a no operation client, its only purpose is to provide an implementation for use in an enviroments that
- * have no proxy to use.
+ * This client implements the interfaces but does nothing.
+ *
+ * It is useful when testing code that needs a ProxyClient, or to configure in
+ * environments that have no caching proxy to talk to, like a local development
+ * setup.
  *
  * @author Gavin Staniforth <gavin@gsdev.me>
  */
-class Noop implements ProxyClient, BanCapable, PurgeCapable, RefreshCapable, TagCapable
+class Noop implements ProxyClient, BanCapable, PurgeCapable, RefreshCapable, TagCapable, ClearCapable
 {
     /**
      * {@inheritdoc}
@@ -70,5 +74,13 @@ class Noop implements ProxyClient, BanCapable, PurgeCapable, RefreshCapable, Tag
     public function flush()
     {
         return 0;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function clear()
+    {
+        return $this;
     }
 }